Fix tests
They still call the live API in the tests, which in the real world
would be bad because it makes the tests slow and dependent upon
another system (i.e. not isolated). If this is here because you're
not sure how to mock out the request, I would suggest focusing on
acquiring that skill.